fix(ui): crash on Color construction from packed sRGB Long
Color(value: ULong) interprets the ULong as Compose's internal packed
representation including a ColorSpace ID in bits 32-37. Passing
parseHexColor()'s ARGB Long converted via .toULong() left those bits
zero/garbage, causing ArrayIndexOutOfBoundsException(length=20;
index=28) inside Color.getColorSpace-impl during text-span layout —
which happened on every recomposition once activeEvent loaded, even if
the card never rendered (Compose composed it speculatively).

Color(color: Long) is the correct overload — it interprets the value
as packed sRGB ARGB and assembles a valid Color internally.

Found by physical-device + simulator crash; both immediately freezing
with the same stack right after Featured tab loaded.

+++ composeApp/src/commonMain/kotlin/com/gallr/app/ui/components/EventPromotionCard.kt
@@ -33,8 +33,8 @@ fun EventPromotionCard(
     onTap: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
-    val brand = parseHexColor(event.brandColor)?.let { Color(it.toULong()) } ?: Color.Black
-    val accent = parseHexColor(event.accentColor)?.let { Color(it.toULong()) }
+    val brand = parseHexColor(event.brandColor)?.let { Color(it) } ?: Color.Black
+    val accent = parseHexColor(event.accentColor)?.let { Color(it) }
 
     val name = event.localizedName(lang)
     val lastToken = Event.nameLastToken(name)
